+++ Scripts/Debugging/LogLevel.cs
@@ -1,0 +1,11 @@
+namespace AeLa.Utilities.Debugging
+{
+	public enum LogLevel
+	{
+		None,
+		Exception,
+		Error,
+		Warning,
+		Info
+	}
+}