+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Util.Random.AlphanumericString
 - AsyncFileWriter utility
 - `Tools/Editor/Open Application.persistentDataPath` menu option
+- LogLevel enum
 
 ### Fixed