+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - DebugFiltered static class for filtered debug logging
 - SceneField
 - WaitForEvent yield instruction
+- `UIFader.OnFadeIn/OutStart` events
 
 ### Fixed