+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SceneField
 - WaitForEvent yield instruction
 - `UIFader.OnFadeIn/OutStart` events
+- `Util.Math.Wrap` for `int` type
+- `Util.Math.Wrap` default overrides for minimum of `0`
 
 ### Changed
 
@@ -31,6 +33,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove NaughtyAttributes dependency in package.json to prevent Unity errors on import
 - Compiler errors in RuntimePlatformFlag.cs in Unity 2022
 - Force set fader state on Fade In/Out when object is disabled
+
+
+### Development
+
+- Added `Util.Math.Wrap` tests
 
 ## [1.3.0] - 2023-10-12