+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] - 2024-05-14
+
+### Fixed
+
+- `SceneField` implicit string conversion throws `NullReferenceException` if the `SceneField` is null
+- `SceneField.RefreshSceneName()` sets `ScenePath` to the scene's name insead of path
+
+### Development
+
+- Added `SceneField` tests
+
 ## [1.4.0] - 2024-05-07
 
 ### Added
@@ -33,7 +44,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove NaughtyAttributes dependency in package.json to prevent Unity errors on import
 - Compiler errors in RuntimePlatformFlag.cs in Unity 2022
 - Force set fader state on Fade In/Out when object is disabled
-
 
 ### Development